+++ src/main/java/module-info.java
@@ -6,4 +6,5 @@ module cleopetra {
 
     exports dk.aau.cs.ds306e18.tournament;
     opens dk.aau.cs.ds306e18.tournament.ui;
+    opens dk.aau.cs.ds306e18.tournament.utility;
 }