+++ changelog.md
@@ -2,6 +2,7 @@
 
 #### Version 1.7.5 - January 2021
 - Updated expected bot pack path and only load bot pack on request from user. - NicEastvillage
+- Bot collection can contain bots with identical names. - NicEastvillage
 
 #### Version 1.7.4 - January 2021
 - New installer that no longer requires users to have Java,