+++ changelog.md
@@ -1,8 +1,9 @@
 # Changelog
 
 #### Unreleased changes
-- Fixed bug that would prevent overlay data to be written on match start. - NicEastvillage
+- Add functionality to create a team for each bot in the bot collection. #114 - NicEastvillage
 - Change radio buttons to checkboxes in RLBot settings tab. - NicEastvillage
+- Fixed bug that would prevent overlay data to be written on match start. - NicEastvillage
 
 #### Version 1.7.5 - January 2021
 - Updated expected bot pack path and only load bot pack on request from user. - NicEastvillage